+++ TabLayout/TabLayout.js
@@ -142,13 +142,29 @@ const TabLayoutBase = kind({
 		/**
 		 * Orientation of the tabs.
 		 *
-		 * Horizontal tabs support a maximum of five tabs.
+		 * Horizontal tabs support a maximum of six tabs.
 		 *
 		 * @type {('horizontal'|'vertical')}
 		 * @default 'vertical'
 		 * @public
 		 */
-		orientation: PropTypes.oneOf(['horizontal', 'vertical'])
+		orientation: PropTypes.oneOf(['horizontal', 'vertical']),
+
+		/**
+		 * Assign a custom size to horizontal tabs.
+		 *
+		 * Tabs in the horizontal orientation automatically stretch to fill the available width.
+		 * Leave this prop blank to use the default auto-sizing behavior.
+		 * Tabs may also be set to a finite width using this property. This accepts numeric pixel
+		 * values. Be mindful of the value you provide as values that are too wide will run off the
+		 * edge of the screen.
+		 *
+		 * Only applies to `orientation="horizontal"` at this time.
+		 *
+		 * @type {Number}
+		 * @public
+		 */
+		tabSize: PropTypes.number
 	},
 
 	defaultProps: {
@@ -189,24 +205,27 @@ const TabLayoutBase = kind({
 			orientation
 		),
 		tabOrientation: ({orientation}) => orientation === 'vertical' ? 'horizontal' : 'vertical',
-		// limit to 5 tabs for horizontal orientation
+		// limit to 6 tabs for horizontal orientation
 		tabs: ({children, orientation}) => {
 			const tabs = React.Children.map(children, (child) => {
 				const {disabled, icon, title} = child.props;
 				return {disabled, icon, title};
 			});
-			return orientation === 'horizontal' && tabs.length > 5 ? tabs.slice(0, 5) : tabs;
+			return orientation === 'horizontal' && tabs.length > 6 ? tabs.slice(0, 6) : tabs;
 		}
 	},
 
-	render: ({children, collapsed, css, dimensions, index, onCollapse, onExpand, onSelect, orientation, tabOrientation, tabs, ...rest}) => {
-		const tabSize = (collapsed ? dimensions.tabs.collapsed : dimensions.tabs.normal);
+	render: ({children, collapsed, css, dimensions, tabSize, index, onCollapse, onExpand, onSelect, orientation, tabOrientation, tabs, ...rest}) => {
+		const tabsSize = (collapsed ? dimensions.tabs.collapsed : dimensions.tabs.normal);
 		const contentSize = (collapsed ? dimensions.content.expanded : dimensions.content.normal);
+		const isVertical = orientation === 'vertical';
+
 		return (
 			<Layout {...rest} orientation={tabOrientation}>
-				<Cell className={css.tabs} size={tabSize}>
+				<Cell className={css.tabs} size={isVertical ? tabsSize : null} shrink={!isVertical}>
 					<TabGroup
-						collapsed={collapsed}
+						collapsed={isVertical ? collapsed : false}
+						tabSize={tabSize}
 						onFocus={onExpand}
 						onFocusTab={onSelect}
 						onSelect={onSelect}
@@ -216,7 +235,7 @@ const TabLayoutBase = kind({
 					/>
 				</Cell>
 				<Cell
-					size={contentSize}
+					size={isVertical ? contentSize : null}
 					className={css.content}
 					component={ViewManager}
 					index={index}