+++ tests/screenshot/apps/components/WizardPanels.js
@@ -1,7 +1,7 @@
 import WizardPanels, {Panel} from '../../../../WizardPanels';
 import React from 'react';
 
-import {withConfig} from './utils';
+import {LongerLoremString, withConfig} from './utils';
 
 const WizardPanelTests = withConfig({
 	wrapper: {
@@ -10,6 +10,10 @@ const WizardPanelTests = withConfig({
 }, [
 	<WizardPanels>
 		<Panel>View 1</Panel>
+		<Panel>View 2</Panel>
+	</WizardPanels>,
+	<WizardPanels>
+		<Panel title="My Title" subtitle={LongerLoremString}>View 1</Panel>
 		<Panel>View 2</Panel>
 	</WizardPanels>,
 	<WizardPanels index={1}>
@@ -22,16 +26,6 @@ const WizardPanelTests = withConfig({
 		<Panel>View 2</Panel>
 		<Panel>View 3</Panel>
 	</WizardPanels>,
-	{
-		component: (
-			<WizardPanels index={1} nextButtonText="nextButtonText" prevButtonText="prevButtonText" title="WizardPanel">
-				<Panel>View 1</Panel>
-				<Panel>View 2</Panel>
-				<Panel>View 3</Panel>
-			</WizardPanels>
-		),
-		locale: 'ar-SA'
-	},
 	// Test unbalanced next/prev button text
 	<WizardPanels index={0} prevButtonText="prevButtonText" title="WizardPanel">
 		<Panel>View 1</Panel>
@@ -42,7 +36,18 @@ const WizardPanelTests = withConfig({
 		<Panel>View 1</Panel>
 		<Panel>View 2</Panel>
 		<Panel>View 3</Panel>
-	</WizardPanels>
+	</WizardPanels>,
+	// RTL
+	{
+		component: (
+			<WizardPanels index={1} nextButtonText="nextButtonText" prevButtonText="prevButtonText" title="WizardPanel">
+				<Panel>View 1</Panel>
+				<Panel>View 2</Panel>
+				<Panel>View 3</Panel>
+			</WizardPanels>
+		),
+		locale: 'ar-SA'
+	}
 ]);
 
 export default WizardPanelTests;